rename NFTMarket to nftMarket
