+++ src/legacy/Routes.tsx
@@ -49,7 +49,7 @@ import PastelUtils from '../common/utils/utils'
 import Creator from '../features/creator'
 import Collector from '../features/collector'
 import Nft from '../features/nft'
-import NFTMarketFeed from '../features/NFTMarket'
+import NFTMarketFeed from '../features/nftMarket'
 import { app } from 'electron'
 
 export type TWalletInfo = {
